push for remote access

+++ ae_on_heads_w_keith/train_sae.py
@@ -29,7 +29,7 @@ def train(encoder :z_sae.AutoEncoder, cfg :z_sae.AutoEncoderConfig, buffer :z_sa
             if i % 100 == 99 and encoder.neurons_reset != None:
                 encoder.re_init_neurons(x - x_reconstruct)
             if i % 100 == 99:
-                encoder.re_init_neurons(x - x_reconstruct)
+                encoder.re_init_neurons(x.float() - x_reconstruct.float())
             loss = encoder.get_loss()
             l2_loss = encoder.l2_loss_cached.mean()
             l1_loss = encoder.l1_loss_cached.mean()
